+++ src/main/java/com/TwoSeaU/BaData/domain/trade/dto/request/SaveGifticonPostRequest.java
@@ -6,8 +6,9 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
@@ -29,7 +30,7 @@ public class SaveGifticonPostRequest {
 
     @NotNull(message = "가격은 필수 작성입니다.")
     @PositiveOrZero(message = "가격은 양수여야 합니다.")
-    private Integer price;
+    private BigDecimal price;
 
     private String comment;
