+++ src/main/java/com/TwoSeaU/BaData/domain/trade/dto/response/GetGifticonDetailResponse.java
@@ -2,6 +2,8 @@ package com.TwoSeaU.BaData.domain.trade.dto.response;
 
 import com.TwoSeaU.BaData.domain.trade.entity.Gifticon;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -13,7 +15,7 @@ public class GetGifticonDetailResponse {
     private Long id;
     private String title;
     private String comment;
-    private Integer price;
+    private BigDecimal price;
     private LocalDate deadLine;
     private String postImage;
     private Boolean isSold;
